update mock in tagged test to return list response
The real tumblr api returns a JSON array response for the /tagged
endpoint, but the mock in the test for the `tagged` method returns a
JSON object. This was causing the test to give a false negative, since
the `tagged` method does not in fact work. Now that the mock returns an
array, the test correctly fails.