fix broken handling of list responses from api
While most endpoints return object responses, the /tagged endpoint
returns an array response. The send_api_request method was handling the
response as though it were always a dict, causing a failure when calling
the `response.get` method on a list. This commit updates the method to
check the type of the response and handle it appropriately.